create common SIMDe header
